+++ python/GafferSceneUI/VisualiserToolUI.py
@@ -34,6 +34,11 @@
 #
 ##########################################################################
 
+import functools
+
+import IECore
+import IECoreScene
+
 import Gaffer
 import GafferUI
 import GafferSceneUI
@@ -52,18 +57,22 @@ Gaffer.Metadata.registerNode(
 	"order", 8,
 	"tool:exclusive", False,
 
+	"toolbarLayout:activator:modeIsColorManualRange", lambda node : node["mode"].getValue() == GafferSceneUI.VisualiserTool.Mode.ColorManualRange,
+
 	plugs = {
 
 		"dataName" : [
 
 			"description",
 			"""
-			Specifies the name of the primitive variable to visualise. The data should
-			be of type float, V2f or V3f.
+			Specifies the name of the primitive variable to visualise. Variables of
+			type int, float, V2f, Color3f or V3f can be visualised.
 			""",
 
 			"toolbarLayout:section", "Bottom",
 			"toolbarLayout:width", 150,
+
+			"plugValueWidget:type", "GafferSceneUI.VisualiserToolUI._DataNameChooser",
 
 		],
 		"opacity" : [
@@ -75,6 +84,34 @@ Gaffer.Metadata.registerNode(
 
 			"toolbarLayout:section", "Bottom",
 			"toolbarLayout:width", 100,
+
+		],
+		"mode" : [
+
+			"description",
+			"""
+			The method for displaying the data.
+
+			Float and integer data are displayed as grayscale. V2f data is displayed with
+			the `x` value for red, `y` value for green and `0` for blue. Vector data is
+			displayed with the `x` value for red, `y` value for green and `z` value for blue.
+			Color data is displayed directly.
+
+			- Auto : Same as `Color (Type Range)`.
+			- Color (Type Range) : Float, integer, V2f and color data is displayed without
+			modification. Vector data is remapped from `[-1, 1]` to `[0, 1]`.
+			- Color (Manual Range) : Values are remapped from the range `[valueMin, valueMax]` to
+			`[0, 1]`.
+			""",
+
+			"preset:Auto", GafferSceneUI.VisualiserTool.Mode.Auto,
+			"preset:Color (Type Range)", GafferSceneUI.VisualiserTool.Mode.ColorTypeRange,
+			"preset:Color (Manual Range)", GafferSceneUI.VisualiserTool.Mode.ColorManualRange,
+
+			"plugValueWidget:type", "GafferUI.PresetsPlugValueWidget",
+
+			"toolbarLayout:section", "Bottom",
+			"toolbarLayout:width", 150,
 
 		],
 		"valueMin" : [
@@ -90,6 +127,8 @@ Gaffer.Metadata.registerNode(
 			"toolbarLayout:section", "Bottom",
 			"toolbarLayout:width", 175,
 
+			"toolbarLayout:visibilityActivator", "modeIsColorManualRange",
+
 		],
 		"valueMax" : [
 
@@ -103,6 +142,8 @@ Gaffer.Metadata.registerNode(
 
 			"toolbarLayout:section", "Bottom",
 			"toolbarLayout:width", 175,
+
+			"toolbarLayout:visibilityActivator", "modeIsColorManualRange",
 
 		],
 		"size": [
@@ -118,3 +159,83 @@ Gaffer.Metadata.registerNode(
 
 	},
 )
+
+class _DataNameChooser( GafferUI.PlugValueWidget ) :
+
+	def __init__( self, plug, **kw ) :
+
+		self.__menuButton = GafferUI.MenuButton(
+			text = plug.getValue(),
+			menu = GafferUI.Menu( Gaffer.WeakMethod( self.__menuDefinition ) )
+		)
+
+		GafferUI.PlugValueWidget.__init__( self, self.__menuButton, plug, **kw )
+
+	def __menuDefinition( self ) :
+
+		menuDefinition = IECore.MenuDefinition()
+
+		node = self.getPlug().node()
+		if not isinstance( node, GafferSceneUI.VisualiserTool ) :
+			return
+		if self.getPlug() != node["dataName"] :
+			return
+
+		scenePlug = node.view()["in"].getInput()
+		scriptNode = node.view().scriptNode()
+		with node.view().context() :
+			selection = GafferSceneUI.ScriptNodeAlgo.getSelectedPaths( scriptNode )
+
+			primVars = set()
+
+			for path in selection.paths() :
+				if not scenePlug.exists( path ) :
+					continue
+
+				primitive = scenePlug.object( path )
+				if not isinstance( primitive, IECoreScene.MeshPrimitive ) :
+					continue
+
+				for v in primitive.keys() :
+					if primitive[v].interpolation not in [
+						IECoreScene.PrimitiveVariable.Interpolation.FaceVarying,
+						IECoreScene.PrimitiveVariable.Interpolation.Uniform,
+						IECoreScene.PrimitiveVariable.Interpolation.Vertex,
+					] :
+						continue
+
+					if not isinstance(
+						primitive[v].data,
+						(
+							IECore.IntVectorData,
+							IECore.FloatVectorData,
+							IECore.V2fVectorData,
+							IECore.Color3fVectorData,
+							IECore.V3fVectorData,
+						)
+					) :
+						continue
+
+					primVars.add( v )
+
+		if len( primVars ) == 0 :
+			menuDefinition.append( "/None Available", { "active" : False } )
+
+		else :
+			for v in reversed( sorted( primVars ) ) :
+				menuDefinition.prepend(
+					"/" + v,
+					{
+						"command" : functools.partial( Gaffer.WeakMethod( self.__setDataName ), v ),
+						"checkBox" : self.getPlug().getValue() == v,
+					}
+				)
+
+		menuDefinition.prepend( "/PrimVarDivider", { "divider" : True, "label" : "Primitive Variables" } )
+
+		return menuDefinition
+
+	def __setDataName( self, value, *unused ) :
+
+		self.getPlug().setValue( value )
+		self.__menuButton.setText( value )